requirements.txt
fastapi
uvicorn
gunicorn
requests
celery
cassandra-driver
dateparser
pandas
requests_html
selenium
bcrypt
secrets